+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Consider new control-plane label.
+
 ## [5.6.1] - 2023-05-03
 
 ### Fixed